FK Fixes & Transaction Rollup

+++ helloworld/views.py
@@ -1,14 +1,15 @@
 ## Forms
-from .forms import CompanyForm
+from .forms import PendingCompanyForm
 from .forms import CategoryForm
 from .forms import SolutionForm
 from .forms import stakeholderGroupsForm
 from .forms import StageForm
 from .forms import ProductGroupForm
-from .forms import ProcessingFocusForm
-from .forms import ExtractionTypeForm
 from .forms import UserRegisterForm
 from .forms import SearchForm
+from .forms import GrowerForm
+from .forms import IndustryForm
+from .forms import StatusForm
 ## Models
 from .models import Company
 from .models import PendingCompany
@@ -17,8 +18,10 @@ from .models import Solution
 from .models import stakeholderGroups
 from .models import Stage
 from .models import ProductGroup
-from .models import ProcessingFocus
-from .models import ExtractionType
+from .models import PendingChanges
+from .models import Grower
+from .models import Industry
+from .models import Status
 ## Django 
 from django.shortcuts import render, redirect
 from django.views.generic import ListView, CreateView
@@ -99,16 +102,36 @@ def company_list(request):
 def companies(request):
     companies = Company.objects.all()
     if request.method == 'POST':
-        form = CompanyForm(request.POST)
+        form = PendingCompanyForm(request.POST)
         if form.is_valid():
-            form.save()
+            company = form.save()
             messages.info(request, 'Company successfully submitted')
+            PendingChanges.objects.create(companyId=company.id, changeType='create')
             return redirect('/companies')  # Redirect to a success page
     else:
-        form = CompanyForm()
+        form = PendingCompanyForm()
         searchForm = SearchForm()
 
     return render(request, 'companies.html', {'form': form, 'companies': companies, 'searchForm': searchForm})
+
+@login_required
+def edit_company(request, id):
+    company = Company.objects.get(id = id)
+    form = PendingCompanyForm(request.POST, instance=company)
+    if request.POST and form.is_valid():
+        company_edit = form.save(commit=False)
+        new_company = PendingCompany()
+        for field in new_company._meta.fields:
+            if not field.primary_key:
+                setattr(new_company, field.name, getattr(company_edit, field.name))
+        new_company.save()
+        messages.info(request, 'Company successfully edited')
+        PendingChanges.objects.create(companyId=new_company.id, changeType='edit', editId=company.id)
+        return redirect('/companies')  # Redirect to a success page
+    else: 
+        form = PendingCompanyForm(instance=company)
+
+    return render(request, 'edit_companies.html', {'form': form, 'company': company})
 
 def view_company(request, id):
     company = Company.objects.get(id = id)
@@ -117,44 +140,70 @@ def view_company(request, id):
 
     return render(request, 'company_view.html', {'company': company, 'obj': obj})
 
-def view_company_pending(request, id):
-    company = PendingCompany.objects.get(id = id)
+def view_company_pending(request, changeType, id):
+    if changeType == 'deletion':
+        company = Company.objects.get(id = id)
+    if changeType == 'create' or changeType == 'edit':
+        company = PendingCompany.objects.get(id = id)    
     obj = model_to_dict(company)
+        
+    return render(request, 'company_view_pending.html', {'company': company, 'obj': obj, 'changeType': changeType})
 
-    return render(request, 'company_view_pending.html', {'company': company, 'obj': obj})
+def view_company_approve(_, changeType, id):
+    if changeType == 'deletion':
+        # Delete Company and Change
+        company = Company.objects.get(id = id)
+        company.delete()
+        change = PendingChanges.objects.get(companyId = id, changeType = changeType)
+        change.delete()
 
-def view_company_approve(request, id):
-    company = PendingCompany.objects.get(id = id)
-    new_company = Company()
-    for field in company._meta.fields:
-        if not field.primary_key:
-            setattr(new_company, field.name, getattr(company, field.name))
-    new_company.save()
-    company.delete()
+        return redirect('/companies')
+    
+    pendingCompany = PendingCompany.objects.get(id = id)
+    change = PendingChanges.objects.get(companyId = id, changeType = changeType)
+    if changeType == 'create':
+        # Create new Company, copy over fields
+        new_company = Company()
+        for field in pendingCompany._meta.fields:
+            if not field.primary_key:
+                setattr(new_company, field.name, getattr(pendingCompany, field.name))
+        new_company.save()
+    if changeType == 'edit':
+        # Find company to edit, copy over fields
+        company = Company.objects.get(id = change.editId)
+        for field in pendingCompany._meta.fields:
+            if not field.primary_key:
+                setattr(company, field.name, getattr(pendingCompany, field.name))
+        company.save()
+
+    # Delete change and pending company
+    change.delete()
+    pendingCompany.delete()
 
     return redirect('/companies')
 
-def view_company_reject(request, id):
-    company = PendingCompany.objects.get(id = id)
-    company.delete()
+def view_company_reject(_, changeType, id):
+    change = PendingChanges.objects.get(companyId = id, changeType = changeType)
+    if changeType == 'create' or changeType == 'edit':
+        company = PendingCompany.objects.get(id=id)
+        company.delete()
+    change.delete()
 
     return redirect('/changes')
-    
-def company_approve(request, id):
-    return
 
 def companies_filtered(request):
     form = SearchForm(request.POST)
     query = form['q']
     companies = Company.objects.filter(Name__contains=query.value())
     
-    form = CompanyForm()
+    form = PendingCompanyForm()
     searchForm = SearchForm()
     return render(request, 'companies.html', {'form': form, 'companies': companies, 'searchForm': searchForm, 'query': query.value()})
 
 def remove_companies(request, id):
-    company = Company.objects.get(id = id)
-    company.delete()
+    # Add deletion to pending changes
+    PendingChanges.objects.create(companyId=id, changeType='deletion')
+    messages.info(request, 'Deletion of Company requested')
     return redirect('/companies')
 
 def export_companies(request):
@@ -184,7 +233,6 @@ def categories(request):
     categories = Category.objects.all()
     if request.method == 'POST':
         form = CategoryForm(request.POST)
-        print(form)
         if form.is_valid():
             form.save()
             return redirect('/categories')  # Redirect to a success page
@@ -207,7 +255,6 @@ def solutions(request):
     solutions = Solution.objects.all()
     if request.method == 'POST':
         form = SolutionForm(request.POST)
-        print(form)
         if form.is_valid():
             form.save()
             return redirect('/solutions')  # Redirect to a success page
@@ -230,7 +277,6 @@ def StakeholderGroups(request):
     groups = stakeholderGroups.objects.all()
     if request.method == 'POST':
         form = stakeholderGroupsForm(request.POST)
-        print(form)
         if form.is_valid():
             form.save()
             return redirect('/stakeholder-groups')  # Redirect to a success page
@@ -253,7 +299,6 @@ def stages(request):
     stages = Stage.objects.all()
     if request.method == 'POST':
         form = StageForm(request.POST)
-        print(form)
         if form.is_valid():
             form.save()
             return redirect('/stages')  # Redirect to a success page
@@ -276,7 +321,6 @@ def productGroups(request):
     groups = ProductGroup.objects.all()
     if request.method == 'POST':
         form = ProductGroupForm(request.POST)
-        print(form)
         if form.is_valid():
             form.save()
             return redirect('/product-groups')  # Redirect to a success page
@@ -290,57 +334,92 @@ def remove_product_groups(request, id):
     group.delete()
     return redirect('/product-groups')
 
-## Processing Focus
-# path('processing-focus/', views.processingFocus, name="processingFocus"),
-# path('remove_focus/<int:id>', views.remove_processing_focus),
+## Status
+# path('status/', views.status, name="status"),
+# path('remove_status/<int:id>', views.remove_status),
 
 @login_required
-def processingFocus(request):
-    focus = ProcessingFocus.objects.all()
-    if request.method == 'POST':
-        form = ProcessingFocusForm(request.POST)
-        print(form)
-        if form.is_valid():
-            form.save()
-            return redirect('/processing-focus')  # Redirect to a success page
-    else:
-        form = ProcessingFocusForm()
-
-    return render(request, 'processingFocus.html', {'form': form, 'processingFocus': focus})
-
-def remove_processing_focus(request, id):
-    focus = ProcessingFocus.objects.get(id = id)
-    focus.delete()
-    return redirect('/processing-focus')
-
-## Extraction Type
-# path('extraction-types/', views.extractionTypes, name="extractionTypes"),
-# path('remove_type/<int:id>', views.remove_extraction_type),
+def status_list(request):
+    status = Status.objects.all()
+    return render(request, 'status.html', {'status': status})
 
 @login_required
-def extractionTypes(request):
-    types = ExtractionType.objects.all()
+def status(request):
+    status = Status.objects.all()
     if request.method == 'POST':
-        form = ExtractionTypeForm(request.POST)
-        print(form)
+        form = StatusForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('/extraction-types')  # Redirect to a success page
+            return redirect('/status')  # Redirect to a success page
     else:
-        form = ExtractionTypeForm()
+        form = StatusForm()
 
-    return render(request, 'extractionTypes.html', {'form': form, 'types': types})
+    return render(request, 'status.html', {'form': form, 'status': status})
 
-def remove_extraction_type(request, id):
-    _type = ExtractionType.objects.get(id = id)
-    _type.delete()
-    return redirect('/extraction-types')
+def remove_status(_, id):
+    status = Status.objects.get(id = id)
+    status.delete()
+    return redirect('/status')
+
+## Grower
+# path('grower/', views.grower, name="grower"),
+# path('remove_grower/<int:id>', views.remove_grower),
+
+@login_required
+def grower_list(request):
+    growers = Grower.objects.all()
+    return render(request, 'grower.html', {'growers': growers})
+
+@login_required
+def grower(request):
+    growers = Grower.objects.all()
+    if request.method == 'POST':
+        form = GrowerForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('/grower')  # Redirect to a success page
+    else:
+        form = GrowerForm()
+
+    return render(request, 'grower.html', {'form': form, 'growers': growers})
+
+def remove_grower(_, id):
+    grower = Grower.objects.get(id = id)
+    grower.delete()
+    return redirect('/grower')
+
+## Industry
+# path('industry/', views.industry, name="industry"),
+# path('remove_industry/<int:id>', views.remove_industry),
+
+@login_required
+def industry_list(request):
+    industries = Industry.objects.all()
+    return render(request, 'industry.html', {'industries': industries})
+
+@login_required
+def industry(request):
+    industries = Industry.objects.all()
+    if request.method == 'POST':
+        form = IndustryForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('/industry')  # Redirect to a success page
+    else:
+        form = IndustryForm()
+
+    return render(request, 'industry.html', {'form': form, 'industries': industries})
+
+def remove_industry(_, id):
+    industry = Industry.objects.get(id = id)
+    industry.delete()
+    return redirect('/industry')
 
 ## Changes
 # path('changes/', views.changes),
 
 @login_required
 def dbChanges(request):
-    changes = PendingCompany.objects.all()
+    changes = PendingChanges.objects.all()
     
-    return render(request, 'companies_pending.html', {'companies': changes})
+    return render(request, 'companies_pending.html', {'changes': changes})
